fix resource test
