+++ tron/api/resource.py
@@ -25,6 +25,7 @@ from twisted.internet import reactor, threads
 from tron import event
 from tron.api import adapter, controller
 from tron.api import requestargs
+from tron.api.async_resource import AsyncResource
 from tron.utils import maybe_decode
 
 log = logging.getLogger(__name__)
@@ -85,50 +86,6 @@ def resource_from_collection(collection, name, child_resource):
     if item is None:
         return resource.NoResource("Cannot find child %s" % name)
     return child_resource(item)
-
-
-class AsyncResource():
-    capacity = 10
-    semaphore = threading.Semaphore(value=capacity)
-    lock = threading.Lock()
-
-    @staticmethod
-    def finish(result, request):
-        request.write(result)
-        request.finish()
-
-    @staticmethod
-    def process(fn, resource, request):
-        with AsyncResource.semaphore:
-            return fn(resource, request)
-
-    @staticmethod
-    def bounded(fn):
-        def wrapper(resource, request):
-            d = threads.deferToThread(
-                AsyncResource.process, fn, resource, request
-            )
-            d.addCallback(AsyncResource.finish, request)
-            d.addErrback(lambda f: f)
-            return server.NOT_DONE_YET
-
-        return wrapper
-
-    @staticmethod
-    def exclusive(fn):
-        def wrapper(*args, **kwargs):
-            # ensures only one exclusive request starts consuming the semaphore
-            with AsyncResource.lock:
-                # this will wait until all bounded requests finished processing
-                for _ in range(AsyncResource.capacity):
-                    AsyncResource.semaphore.acquire()
-                try:
-                    return fn(*args, **kwargs)
-                finally:
-                    for _ in range(AsyncResource.capacity):
-                        AsyncResource.semaphore.release()
-
-        return wrapper
 
 
 class ActionRunResource(resource.Resource):
